+++ src/Resources/contao/dca/tl_newsletter_items.php
@@ -1,0 +1,334 @@
+<?php
+
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2005-2015 Leo Feyer
+ *
+ * @package   Elo
+ * @author    Frank Hoppe
+ * @license   GNU/LPGL
+ * @copyright Frank Hoppe 2016
+ */
+
+/**
+ * Table tl_newsletter_items
+ */
+$GLOBALS['TL_DCA']['tl_newsletter_items'] = array
+(
+
+	// Config
+	'config' => array
+	(
+		'dataContainer'             => 'Table',
+		'ptable'                    => 'tl_newsletter',
+		'enableVersioning'          => true,
+		'sql' => array
+		(
+			'keys' => array
+			(
+				'id'                            => 'primary',
+				'pid'                           => 'index',
+			)
+		)
+	),
+
+	// List
+	'list' => array
+	(
+		'sorting' => array
+		(
+			'mode'                    => 4,
+			'fields'                  => array('sorting'),
+			'flag'                    => 1,
+			'headerFields'            => array('subject'), 
+			'panelLayout'             => 'sort,filter;search,limit',
+			'child_record_callback'   => array('tl_newsletter_items', 'listContent'),
+		),
+		'global_operations' => array
+		(
+			'all' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['MSC']['all'],
+				'href'                => 'act=select',
+				'class'               => 'header_edit_all',
+				'attributes'          => 'onclick="Backend.getScrollOffset();" accesskey="e"'
+			)
+		),
+		'operations' => array
+		(
+			'edit' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_newsletter_items']['edit'],
+				'href'                => 'act=edit',
+				'icon'                => 'edit.gif'
+			),
+			'copy' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_newsletter_items']['copy'],
+				'href'                => 'act=copy',
+				'icon'                => 'copy.gif'
+			),
+			'cut' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_newsletter_items']['cut'],
+				'href'                => 'act=paste&amp;mode=cu',
+				'icon'                => 'cut.gif'
+			),
+			'delete' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_newsletter_items']['delete'],
+				'href'                => 'act=delete',
+				'icon'                => 'delete.gif',
+				'attributes'          => 'onclick="if(!confirm(\'' . $GLOBALS['TL_LANG']['MSC']['deleteConfirm'] . '\'))return false;Backend.getScrollOffset()"'
+			),
+			'toggle' => array
+			(
+				'href'                => 'act=toggle&amp;field=invisible',
+				'icon'                => 'visible.svg',
+				'button_callback'     => array('tl_newsletter_items', 'toggleIcon')
+			),
+			//'toggle' => array
+			//(
+			//	'label'                => &$GLOBALS['TL_LANG']['tl_newsletter_items']['toggle'],
+			//	'attributes'           => 'onclick="Backend.getScrollOffset()"',
+			//	'haste_ajax_operation' => array
+			//	(
+			//		'field'            => 'invisible',
+			//		'options'          => array
+			//		(
+			//			array('value' => '1', 'icon' => 'invisible.svg'),
+			//			array('value' => '', 'icon' => 'visible.svg'),
+			//		),
+			//	),
+			//),
+			'show' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_newsletter_items']['show'],
+				'href'                => 'act=show',
+				'icon'                => 'show.gif'
+			)
+		)
+	),
+
+	// Palettes
+	'palettes' => array
+	(
+		'__selector__'                => array('useImage', 'overwriteMeta'),
+		'default'                     => '{title_legend},headline;{image_legend},useImage;{text_legend},text;{invisible_legend:hide},invisible'
+	),
+
+	// Subpalettes
+	'subpalettes' => array
+	(
+		'useImage'                    => 'singleSRC,overwriteMeta', // size entfernt, da im Newsletter nur Oiginalbilder verwendet werden sollen
+		'overwriteMeta'               => 'alt,imageTitle,imageUrl,caption'
+	),
+
+	// Fields
+	'fields' => array
+	(
+		'id' => array
+		(
+			'sql'                     => "int(10) unsigned NOT NULL auto_increment"
+		),
+		'pid' => array
+		(
+			'sql'                     => "int(10) unsigned NOT NULL default '0'"
+		),
+		'sorting' => array
+		(
+			'sql'                     => "int(10) unsigned NOT NULL default 0"
+		),
+		'tstamp' => array
+		(
+			'sql'                     => "int(10) unsigned NOT NULL default '0'"
+		),
+		'headline' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_newsletter_items']['headline'],
+			'exclude'                 => true,
+			'search'                  => true,
+			'inputType'               => 'inputUnit',
+			'options'                 => array('h1', 'h2', 'h3', 'h4', 'h5', 'h6'),
+			'eval'                    => array('maxlength'=>200, 'tl_class'=>'w50 clr'),
+			'sql'                     => "varchar(255) NOT NULL default 'a:2:{s:5:\"value\";s:0:\"\";s:4:\"unit\";s:2:\"h2\";}'"
+		),
+		'text' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_newsletter_items']['text'],
+			'exclude'                 => true,
+			'search'                  => true,
+			'inputType'               => 'textarea',
+			'eval'                    => array('mandatory'=>true, 'rte'=>'tinyMCE', 'helpwizard'=>true),
+			'explanation'             => 'insertTags',
+			'sql'                     => "mediumtext NULL"
+		),
+		'overwriteMeta' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_newsletter_items']['overwriteMeta'],
+			'exclude'                 => true,
+			'inputType'               => 'checkbox',
+			'eval'                    => array('submitOnChange'=>true, 'tl_class'=>'w50 clr'),
+			'sql'                     => "char(1) COLLATE ascii_bin NOT NULL default ''"
+		),
+		'singleSRC' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_newsletter_items']['singleSRC'],
+			'exclude'                 => true,
+			'inputType'               => 'fileTree',
+			'eval'                    => array
+			(
+				'filesOnly'           => true, 
+				'fieldType'           => 'radio',
+				'extensions'          => '%contao.image.valid_extensions%', 
+				'mandatory'           => true, 
+				'tl_class'            => 'clr'
+			),
+			'sql'                     => "binary(16) NULL"
+		),
+		'alt' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_newsletter_items']['alt'],
+			'exclude'                 => true,
+			'search'                  => true,
+			'inputType'               => 'text',
+			'eval'                    => array('maxlength'=>255, 'tl_class'=>'w50'),
+			'sql'                     => "varchar(255) NOT NULL default ''"
+		),
+		'imageTitle' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_newsletter_items']['imageTitle'],
+			'exclude'                 => true,
+			'search'                  => true,
+			'inputType'               => 'text',
+			'eval'                    => array('maxlength'=>255, 'tl_class'=>'w50'),
+			'sql'                     => "varchar(255) NOT NULL default ''"
+		),
+		'size' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['MSC']['imgSize'],
+			'exclude'                 => true,
+			'inputType'               => 'imageSize',
+			'reference'               => &$GLOBALS['TL_LANG']['MSC'],
+			'eval'                    => array
+			(
+				'rgxp'                => 'natural', 
+				'includeBlankOption'  => true, 
+				'nospace'             => true, 
+				'helpwizard'          => true, 
+				'tl_class'            => 'w50'
+			),
+			'options_callback'        => static function ()
+			{
+				return \System::getContainer()->get('contao.image.image_sizes')->getOptionsForUser(BackendUser::getInstance());
+			},
+			'sql'                     => "varchar(128) COLLATE ascii_bin NOT NULL default ''"
+		),
+		'imageUrl' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_newsletter_items']['imageUrl'],
+			'exclude'                 => true,
+			'search'                  => true,
+			'inputType'               => 'text',
+			'eval'                    => array('rgxp'=>'url', 'decodeEntities'=>true, 'maxlength'=>2048, 'dcaPicker'=>true, 'tl_class'=>'w50'),
+			'sql'                     => "text NULL"
+		),
+		'caption' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_newsletter_items']['caption'],
+			'exclude'                 => true,
+			'search'                  => true,
+			'inputType'               => 'text',
+			'eval'                    => array('maxlength'=>255, 'allowHtml'=>true, 'tl_class'=>'w50'),
+			'sql'                     => "varchar(255) NOT NULL default ''"
+		),
+		'useImage' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_newsletter_items']['useImage'],
+			'exclude'                 => true,
+			'inputType'               => 'checkbox',
+			'eval'                    => array('submitOnChange'=>true),
+			'sql'                     => "char(1) COLLATE ascii_bin NOT NULL default ''"
+		),
+		'invisible' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_newsletter_items']['invisible'],
+			'exclude'                 => true,
+			'toggle'                  => true,
+			'filter'                  => true,
+			'inputType'               => 'checkbox',
+			'sql'                     => "char(1) COLLATE ascii_bin NOT NULL default ''"
+		),
+	)
+);
+
+/**
+ * Provide miscellaneous methods that are used by the data configuration array
+ */
+class tl_newsletter_items extends Backend
+{
+	 
+	/**
+	 * Import the back end user object
+	 */
+	public function __construct()
+	{
+		parent::__construct();
+		$this->import('BackendUser', 'User');
+	}
+
+	/**
+	 * Generiere eine Zeile als HTML
+	 * @param array
+	 * @return string
+	 */
+	public function listContent($arrRow)
+	{
+		$key = $arrRow['invisible'] ? 'unpublished' : 'published';
+		$headline = unserialize($arrRow['headline']);
+		$titel = $headline['value'] ? $headline['value'] : '- ohne Überschrift -';
+		$class = 'limit_height';
+
+		// Limit the element's height
+		if (!\Config::get('doNotCollapse'))
+		{
+			$class .= ' h40';
+		}
+		
+		$content = $arrRow['text'];
+
+		return '
+<div class="cte_type ' . $key . '">' . $titel . '</div>
+<div class="' . trim($class) . '">
+' . \StringUtil::insertTagToSrc($content) . '
+</div>' . "\n";
+	}
+
+	/**
+	 * Return the "toggle visibility" button
+	 *
+	 * @param array  $row
+	 * @param string $href
+	 * @param string $label
+	 * @param string $title
+	 * @param string $icon
+	 * @param string $attributes
+	 *
+	 * @return string
+	 */
+	public function toggleIcon($row, $href, $label, $title, $icon, $attributes)
+	{
+		$security = \System::getContainer()->get('security.helper');
+
+		$href .= '&amp;id=' . $row['id'];
+
+		if ($row['invisible'])
+		{
+			$icon = 'invisible.svg';
+		}
+
+		return '<a href="' . $this->addToUrl($href) . '" title="' . \StringUtil::specialchars($title) . '" onclick="Backend.getScrollOffset();return AjaxRequest.toggleField(this,true)">' . Image::getHtml($icon, $label, 'data-icon="' . Image::getPath('visible.svg') . '" data-icon-disabled="' . \Image::getPath('invisible.svg') . '" data-state="' . ($row['invisible'] ? 0 : 1) . '"') . '</a> ';
+	}
+
+}