+++ src/Resources/contao/languages/de/tl_newsletter.php
@@ -1,0 +1,7 @@
+<?php
+
+/**
+ * Buttons für Operationen
+ */
+
+$GLOBALS['TL_LANG']['tl_newsletter']['editContent'] = array('Inhaltselemente bearbeiten', 'Inhaltselemente des Newsletters %s bearbeiten');