+++ src/Resources/contao/languages/de/tl_newsletter_items.php
@@ -1,0 +1,36 @@
+<?php
+
+/**
+ * Backend-Modul: Übersetzungen im Eingabeformular
+ */
+$GLOBALS['TL_LANG']['tl_newsletter_items']['title_legend'] = 'Überschrift';
+$GLOBALS['TL_LANG']['tl_newsletter_items']['headline'] = array('Überschrift', 'Hier können Sie dem Eintrag eine Überschrift hinzufügen.');
+
+$GLOBALS['TL_LANG']['tl_newsletter_items']['text_legend'] = 'Text';
+$GLOBALS['TL_LANG']['tl_newsletter_items']['text'] = array('Text', 'Sie können HTML-Tags verwenden, um den Text zu formatieren.');
+
+$GLOBALS['TL_LANG']['tl_newsletter_items']['image_legend'] = 'Bildeinstellungen';
+$GLOBALS['TL_LANG']['tl_newsletter_items']['overwriteMeta'] = array('Metadaten überschreiben', 'Die Metadaten aus der Dateiverwaltung überschreiben.');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['singleSRC'] = array('Bild', 'Bitte wählen Sie ein Bild aus der Dateiübersicht.');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['alt'] = array('Alternativer Text', 'Hier können Sie einen alternativen Text für das Bild eingeben (&lt;em&gt;alt&lt;/em&gt;-Attribut).');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['imageTitle'] = array('Bildtitel', 'Hier können Sie den Titel des Bildes eingeben (&lt;em&gt;title&lt;/em&gt;-Attribut).');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['imageUrl'] = array('Bildlink-Adresse', 'Eine eigene Bildlink-Adresse.');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['caption'] = array('Bildunterschrift', 'Hier können Sie einen kurzen Text eingeben, der unterhalb des Bildes angezeigt wird.');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['useImage'] = array('Ein Bild verwenden', 'Ein Bild verwenden. Es wird oberhalb des Textes angezeigt. Bitte verwenden Sie nur Bilder in Originalauflösung, die für den Newsletter aufbereitet wurden!');
+
+$GLOBALS['TL_LANG']['tl_newsletter_items']['invisible_legend'] = 'Sichtbarkeit';
+$GLOBALS['TL_LANG']['tl_newsletter_items']['invisible'] = array('Unsichtbar', 'Den Eintrag im Newsletter nicht anzeigen.');
+
+$GLOBALS['TL_LANG']['tl_newsletter_items']['headline'] = array('Überschrift', 'Überschrift des Eintrags');
+
+/**
+ * Buttons für Operationen
+ */
+
+$GLOBALS['TL_LANG']['tl_newsletter_items']['new'] = array('Neuer Eintrag', 'Neuen Eintrag anlegen');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['edit'] = array('Eintrag bearbeiten', 'Eintrag %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['copy'] = array('Eintrag kopieren', 'Eintrag %s kopieren');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['cut'] = array('Eintrag duplizieren', 'Eintrag %s duplizieren');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['delete'] = array('Eintrag löschen', 'Eintrag %s löschen');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['toggle'] = array('Eintrag aktivieren/deaktivieren', 'Eintrag %s aktivieren/deaktivieren');
+$GLOBALS['TL_LANG']['tl_newsletter_items']['show'] = array('Eintragsdetails anzeigen', 'Details des Eintrags %s anzeigen');