+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # DSB-Newsletter Changelog
 
+## Version 0.1.0 (2024-11-14)
+
+* Erweiterung ausgebaut
+* Change: Inhaltselemente nach unterhalb tl_newsletter verschoben
+* Add: Ausbau der Inserttag-Ersetzung mittels {{newsletter::id}}
+
 ## Version 0.0.1 (2024-10-29)
 
 * Initialversion als Contao-4-Bundle